Make sure <CR> does not do anything except on regular buffers
Otherwise, it can result in strange artifacts in quickfix, and in the
cmdline-window mode.  Any buffer with a buffer-local '&buftype' that's
an empty string, is a regular file being edited, where we would like to
be enabled.  Other buftypes, there's no reason for stabs to be enabled
on CR (mostly these are file types which are ephemeral read-only
buffers, cmdline-window mode is really still line-based and enter
modifications don't make sense).  The only one maybe is scratch buffer,
but <CR> would still be usable there and it's better than maintaining a
list of exceptions since just empty buftype solves 99% of use cases